+++ Pagination/Pagination.php
@@ -48,7 +48,11 @@ class Pagination extends AbstractPagination
         if (!$this->renderer instanceof Closure) {
             $output = 'add a renderer in order to render a template';
         } else {
-            $output = call_user_func($this->renderer, $data);
+            try {
+                $output = call_user_func($this->renderer, $data);
+            } catch (\Exception $e) {
+                return $e->getMessage();
+            }
         }
 
         return $output;